admin side completed

+++ lib/view/screens/comment/widget/comment_card.dart
@@ -2,24 +2,24 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:intl/intl.dart';
-import 'package:react_messenger/controller/user_controller.dart';
+import '../../../../controller/user_controller.dart';
 import '../../../../services/firestore_methods.dart';
 import '../../../../widgets/widgets.dart';
 
 class CommentCard extends StatelessWidget {
+  CommentCard({super.key, required this.commentSnapShot, this.postId});
   final QueryDocumentSnapshot<Map<String, dynamic>> commentSnapShot;
   final String? postId;
   final UserController userController = Get.put(UserController());
-  CommentCard({super.key, required this.commentSnapShot, this.postId});
 
   @override
   Widget build(BuildContext context) {
     return Container(
       padding: const EdgeInsets.symmetric(vertical: 18, horizontal: 16),
       child: Row(
-        children: [
+        children: <Widget>[
           CircleAvatarWidget(
-            networkImagePath: commentSnapShot['profilePic'],
+            networkImagePath: commentSnapShot['profilePic'] as String,
             radius: 18,
           ),
           Expanded(
@@ -28,12 +28,12 @@ class CommentCard extends StatelessWidget {
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 mainAxisAlignment: MainAxisAlignment.center,
-                children: [
+                children: <Widget>[
                   RichText(
                     text: TextSpan(
-                      children: [
+                      children: <InlineSpan>[
                         TextSpan(
-                          text: commentSnapShot['name'],
+                          text: commentSnapShot['name'] as String,
                           style: const TextStyle(fontWeight: FontWeight.bold),
                         ),
                         TextSpan(
@@ -47,27 +47,28 @@ class CommentCard extends StatelessWidget {
             ),
           ),
           Column(
-            children: [
-              commentSnapShot['uid'] == userController.userData.value!.uid
-                  ? Column(
-                      children: [
-                        IconButton(
-                          onPressed: () async {
-                            await FirestoreMethods().deleteComment(
-                              postId!,
-                              commentSnapShot['commentId'],
-                            );
-                          },
-                          icon: const Icon(Icons.delete),
-                        ),
-                      ],
-                    )
-                  : const SizedBox(),
+            children: <Widget>[
+              if (commentSnapShot['uid'] == userController.userData.value!.uid)
+                Column(
+                  children: <Widget>[
+                    IconButton(
+                      onPressed: () async {
+                        await FirestoreMethods().deleteComment(
+                          postId!,
+                          commentSnapShot['commentId'] as String,
+                        );
+                      },
+                      icon: const Icon(Icons.delete),
+                    ),
+                  ],
+                )
+              else
+                const SizedBox(),
               Padding(
                 padding: const EdgeInsets.only(top: 4),
                 child: Text(
-                  DateFormat.yMMMd()
-                      .format(commentSnapShot['datePublished'].toDate()),
+                  DateFormat.yMMMd().format(
+                      (commentSnapShot['datePublished'] as Timestamp).toDate()),
                   style: const TextStyle(fontSize: 7),
                 ),
               ),
